Add merge-global-schema script
In fact this is not necessary for merge, and
dumps output to a separate file. But running
the script against the old Jurism and new Zotero
versions of the schema will issue chatter flagging
the changes made. That info is useful in making
adjustments to the Jurism patch file read in by
the hacked update-schema script that generates
the drop-in replacement schema-jurism.json.