Support removal of Zotero CSL mappings in schema-update
Zotero 6.0.20 calls the master branch of zotero-schema
in .gitmodules. At the time of the Jurism update, merging
the master head of zotero-schema to its Jurism fork produced
a working schema for the build. The z6-update branch of
zotero-schema seems not to be compatible, as it includes
fields for which no label properties exist in the 6.0.20
locales.

One problem with the master head schema, however, was the
mapping of Zotero fields to CSL event-title rather than event.
Both are valid CSL variable, but only the latter is used
in published styles. To make the change requires complete
removal of the event-title mapping from the schema. That's
implemented by treating a completely empty field mapping
array in schema-jurism-patch.json as a removal request.